+++ keccak256/src/circuit.rs
@@ -1,4 +1,4 @@
-pub mod padding;
+pub mod word_builder;
 
 pub const MAX_INPUT_BYTES: usize = MAX_INPUT_WORDS * BYTES_PER_WORD;
 pub const MAX_INPUT_WORDS: usize = MAX_PERM_ROUNDS * NEXT_INPUTS_WORDS;
