+++ keccak256/src/circuit/word_builder.rs
